+++ src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2019-2022 Artem Vorotnikov and Monero Rust Contributors
+// Copyright 2019-2023 Artem Vorotnikov and Monero Rust Contributors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -1359,6 +1359,35 @@ pub async fn get_transfer_from_str(
         let minor = version.version - (major << 16);
 
         Ok((u16::try_from(major)?, u16::try_from(minor)?))
+    }
+
+    /// Returns an attribute as a string or an error when there is no attribute for the given key
+    pub async fn get_attribute(&self, key: String) -> anyhow::Result<String> {
+        let params = empty().chain(once(("key", key.into())));
+
+        #[derive(Deserialize)]
+        struct Rsp {
+            value: String,
+        }
+
+        Ok(self
+            .inner
+            .request::<Rsp>("get_attribute", RpcParams::map(params))
+            .await?
+            .value)
+    }
+
+    /// Set an arbitrary attribute which is saved in the wallet
+    pub async fn set_attribute(&self, key: String, value: String) -> anyhow::Result<()> {
+        let params = empty()
+            .chain(once(("key", key.into())))
+            .chain(once(("value", value.into())));
+
+        self.inner
+            .request::<IgnoredAny>("set_attribute", RpcParams::map(params))
+            .await?;
+
+        Ok(())
     }
 }