+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-09-15
+
+### Added
+
+- Add `sign` and `verify` methods by @silverpill ([#105](https://github.com/monero-rs/monero-rpc-rs/pull/105))
+- Add `get_attribute` and `set_attribute` methods by @refactor-ring ([#108](https://github.com/monero-rs/monero-rpc-rs/pull/108))
+
+### Changed
+
+- Bump MSRV to `1.63.0`
+- Run test suite against Monero node and wallet `0.18.1.2` and `0.18.2.2`
+
 ## [0.3.2] - 2022-12-13
 
 ### Removed
@@ -129,7 +141,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Daemon RPC methods:
   - `get_transactions`
 
-[Unreleased]: https://github.com/monero-rs/monero-rpc-rs/compare/v0.3.2...HEAD
+[Unreleased]: https://github.com/monero-rs/monero-rpc-rs/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/monero-rs/monero-rpc-rs/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/monero-rs/monero-rpc-rs/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/monero-rs/monero-rpc-rs/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/monero-rs/monero-rpc-rs/compare/v0.2.0...v0.3.0